+++ frontend/src/pages/CourseSelector/courseDescription/Collapsible.jsx
